requirements.txt
langchain==0.3.18 
langchain-community==0.3.17 
langchain-core==0.3.45
langchain-google-genai
python-dotenv
langgraph==0.3.21
grandalf